+++ Source/Lib/Encoder/Codec/EbMotionEstimation.h
@@ -357,16 +357,16 @@ extern "C" {
         uint32_t *p_best_sad_16x16,
         uint32_t *p_best_mv8x8,
         uint32_t *p_best_mv16x16,
-        uint32_t p_eight_sad16x16[16][8],
-        uint32_t p_eight_sad8x8[64][8]);
+        uint16_t p_eight_sad16x16[16][8],
+        uint16_t p_eight_sad8x8[64][8]);
 
     /****************************************************
     Calculate SAD for Rect H, V and H4, V4 partitions
     and update its Motion info if the result SAD is better
     ****************************************************/
     extern void ext_eigth_sad_calculation_nsq_c(
-        uint32_t p_sad8x8[64][8],
-        uint32_t p_sad16x16[16][8],
+        uint16_t p_sad8x8[64][8],
+        uint16_t p_sad16x16[16][8],
         uint32_t p_sad32x32[4][8],
         uint32_t *p_best_sad_64x32,
         uint32_t *p_best_mv64x32,
@@ -396,7 +396,7 @@ extern "C" {
     the best SAD+MV
     *******************************************/
     extern void ext_eight_sad_calculation_32x32_64x64_c(
-        uint32_t p_sad16x16[16][8],
+        uint16_t p_sad16x16[16][8],
         uint32_t *p_best_sad_32x32,
         uint32_t *p_best_sad_64x64,
         uint32_t *p_best_mv32x32,
